desloppify: structural review batch 2 — 6 findings fixed
- Split `AccountState`/`AccountRegistry` and their inherent impls out of
  boot/registry.rs into a new `boot/account_state.rs` module. registry.rs
  now contains only the build pipeline. The 21 importers were updated to
  reference the data types via `crate::boot::account_state::*` so the
  recompile blast radius shrinks when boot logic changes.
- Lift `PostureContext` and `rimap_error_to_breaker_reason` into
  `mcp/posture_context.rs`. Removes the bidirectional file cycle where
  dispatch.rs added inherent impls to ImapMcpServer (defined in
  server.rs) and server.rs imported PostureContext from dispatch.rs.
- Flatten `crates/rimap-server/src/shim/mod.rs` to `shim.rs` (single-file
  directory removed).
- Move `tools/{validation,fetch_by_uid}.rs` into `tools/common/` so the
  tool partition no longer mixes capability folders with flat helpers.
  All call sites use `crate::tools::common::{validation,fetch_by_uid}`.
- Apply `#[non_exhaustive]` uniformly to every public Meta/Entry/Info
  response struct under `tools/`. Document the convention on
  `mcp::response::ToolResponse`.
