+++ crates/rimap-server/src/boot/mod.rs
@@ -1,5 +1,6 @@
 //! Bootstrap modules: logging init, account registry build, audit writer init.
 
+pub mod account_state;
 pub mod audit_init;
 pub mod discovery;
 pub mod logging;
